requirements.txt
Django==5.2
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2
pytesseract~=0.3.13
pillow~=11.2.1
gTTS~=2.5.4
PyPDF2~=3.0.1
python-docx~=1.1.2
pyttsx3~=2.98
pdf2image~=1.17.0